2019-03-13  Benjamin De Kosnik  <bdekoz@mozilla.com>
	* src/a60-svg-base.h (transform::scale): Use floating point args.
	* src/moz-telemetry-x-svg-radial.h: Insert glyph with diameter size.
	* src/moz-telemetry-x-svg.h (insert_svg_at_center): Scale.

+++ src/a60-svg-base.h
@@ -822,7 +822,7 @@ namespace k = constants;
     }
 
     static string
-    scale(int factor)
+    scale(double factor)
     {
       std::ostringstream stream;
       stream << "scale(" << factor << ")";
